Added Support for OIDC- and SAML-Login to website. Implemented package setup file. Added OIDC-Login to Client.
The website now supports OIDC- and SAML-Login under /accounts/login.

As the secrets are in a folder of their own, they must be considered a package. To use a package however, you need to build and install it, which is where setup.py and pip install come into play.

The client application now supports OIDC-Login, yet it is not flawless. SAML-Login does not work, as the current structure of the client resembles an Enhanced Client and is not a website or doesn't emulate a website. There is an entirely different SAML Protocol Flow for Authentication when it comes to ECs, so the SAML-Login has to be rewritten to test the Webbrowser Protocol Flow. This does not mean, that the EC-Flow is irrelevant to this thesis.