Added first iteration of log analyzation
Nothing fancy just yet. I didn't test this, I just pushed it to show my progress. This should read one of my OIDC-Logs and properly put it into a dictionary. I plan on transforming that dictionary into a json and interpret said json with matplotlib and numpy if that is even possible. We'll see.