Refactored analyze.py and made it more DRY
Just moved some common functionality out of the bigger functions. There may be a way to make this even more DRY, but for now this is good enough. Also noticed that checking if we already serialized a log is unnecessary in this implementation due to the way fetch_and_store_log is implemented. At a future date this may be implemented correctly. Next up is data interpretation and visualization.