Moved log processing out of analyze.py. Added autonomous log processing without prompt. Added automatic log clearing on serialization.
This basically does it for log processing. It would be good practice to check if the return of scp actually contains data before I serialize the log, but as this serialization is only called once I actually did a test it should be fine. I also got the automatic log processing to work. This was done by using ssh-agent and its ability to save ssh keyphrases so they do not always have to be typed in. The automatic closing of the ssh-agent service cannot be done by the script though, I may be able to figure something out for this later on if necessary.