Faulty:] Added resource monitoring to the IdP. Added processing of IdP resource monitor record. Fixed bug in user_client where the KcAdmin-Session was initialized before the IdP was shut down.
This is a big commit and one I have been working some days on. It is marked as faulty, because in its current state it doesn't do what it should do and the things it can do it doesn't do correctly. The processing of the IdP record is faulty for example.

The really big reason this is pushed is to commemorate the current solution as it will not be used going further. There was a significant issue with using systemd-cgtop that I noticed a few days ago as well, but chose to ignore for now. The issue is, that systemd-cgtop is not capable of measuring I/O operations etc. without more than one run of the command, as its basically calculating current_total_io - last_total_io = io_change and shows io_change. Its the same issue as to why only CPU time can be measured with systemd-cgtop in one run. Now that shouldn't really be a problem, after all the --batch option exists fort a reason, so just run it twice or rather 10 times every time. The output is conveniently collected and can be written to a file as well. The problem with this approach is, that this would not allow me to put a timestamp to every measurement but the first of the batch, which means I would have to approximate the timestamps for the other measurements, or just accept that I do not have timestamps for them. None of this is an option. systemd-cgtop also doesn't have an option to yield the timestamps it uses to calculate the CPU%, which in turn means systemd-cgtop is out of the picture.

The question is whats next. Whats next is reading the .stat files of the important cgroups at the IdP every second, log them and download this log. This allows me to timestamp every measurement. I have to be careful here to not get lost in formatting early on, as formatting itself costs time, which in turn needs to be accounted for.