Rewritten resource monitor.
This is the working version of the resource monitor. It reads from the .stat files of the relevant controllers from the relevant cgroups. The record is transformed into a dictionary which can be stored as a json-File and interpreted later on. Next up are some changes to the analyze.py file and then I'm done.