Added resource monitor log storage. Added new measurements to log transformations. Fixed clear_log_on_server. Fixed poll_websites_before_eval. Added conditional evaluation based on the success of poll_websites_before_eval.
This is the final patch for the measurements. There was a bug in the code, that allowed an evaluation to fail even after polling a website. To counteract this, the TimeoutException is used as an indicator to re-poll the website for a maximum of 3 tries. If all tries are used the system is reset and the polling is tried again. If that fails 3 times the evaluation cycle failed. This is recorded and analysis is postponed.

The last thing to do is add some analysis code, which should be quickly doable.