Added plotting for resources to analyze. Fixed plotting for eval in analyze. Fixed how timestamps are stored in log_processor.
I noticed that the timestamps are stored as a string in log_processor. This was unintentional. They should be stored as integers now. I also added plotting capabilities for resources to analyze.py. While its still not fully automatic I believe it is good enough. I also fixed how the graph for the evaluation data is displayed. It should have the Users on the x axis now instead of the Evaluation.